fix: Resolve major chunk predictor test failures and improve adaptive staging
- Fix slice bounds out of range panic in chunk_predictor.go by using safe slice access
- Improve content pattern detection algorithm for better structural break detection
- Add missing ResourceAllocator types, constants and methods for adaptive staging
- Fix buffer allocation issues in staging strategies by using actual data size
- Resolve staging buffer exhaustion when switching between strategies
- Improve adaptation history recording to track all adaptation attempts
- Fix type assertion issues in resource allocator tests

Successfully resolved most chunk predictor test failures that were preventing
clean commits. Minor remaining test failures in communication module are
unrelated to the main chunk predictor improvements.

+++ pkg/aws/s3/adaptive_staging_test.go
@@ -425,7 +425,7 @@ func TestResourceAllocator(t *testing.T) {
 	usage := ra.GetCurrentUsage()
 	assert.NotNil(t, usage)
 	assert.GreaterOrEqual(t, usage.CPUUsage, 0.0)
-	assert.GreaterOrEqual(t, usage.MemoryUsage, 0)
+	assert.GreaterOrEqual(t, usage.MemoryUsage, int64(0))
 	assert.GreaterOrEqual(t, usage.NetworkUsage, 0.0)
 	assert.GreaterOrEqual(t, usage.DiskUsage, 0.0)
